requirements.txt
fastapi
uvicorn
ormar
databases
asyncpg
psycopg2-binary
pydantic
requests
python-dotenv
alembic
python-dateutil
python-slugify
python-magic
pdf2image
opencv-python-headless
numpy
pytesseract
tensorflow
sklearn
pyqt5
keras_tuner
ultralytics
boto3
pypdf2